op-reth(cross-unsafe): split into state and source submodules
Move the in-memory cache data structures (CrossUnsafeState, CachedBlock,
SourceRef) into cross_unsafe/state.rs and the source-chain RPC clients
(SourceLogClients, SourceLogClient) into cross_unsafe/source.rs, leaving
cross_unsafe.rs as the public API + orchestration. No behavior change.
