+++ rust/op-reth/crates/rpc/src/cross_unsafe.rs
@@ -11,15 +11,16 @@
 //! forward toward local latest validating new blocks.
 //!
 //! Concurrency: a single mutex protects the cache and is held for the duration of a refresh,
-//! including the source-chain RPC calls it makes. Each source RPC is bounded by
-//! [`SOURCE_RPC_TIMEOUT`], so a slow source cannot stall the endpoint indefinitely, and the
-//! background poller keeps the cache warm so user calls usually have little work to do.
+//! including the source-chain RPC calls it makes. Each source RPC is bounded by a per-request
+//! timeout, so a slow source cannot stall the endpoint indefinitely, and the background poller
+//! keeps the cache warm so user calls usually have little work to do.
+//!
+//! The cache data structures live in the `state` submodule and the source-chain RPC clients in the
+//! `source` submodule.
 
 use alloy_consensus::{BlockHeader, TxReceipt};
 use alloy_eips::BlockHashOrNumber;
 use alloy_primitives::{B256, U64, keccak256};
-use alloy_rpc_client::{ClientBuilder, RpcClient};
-use alloy_rpc_types_eth::{Block as RpcBlock, Filter, Log as RpcLog};
 use jsonrpsee::proc_macros::rpc;
 use jsonrpsee_core::{RpcResult, async_trait};
 use kona_interop::{MESSAGE_EXPIRY_WINDOW, RawMessagePayload, parse_log_to_executing_message};
@@ -27,15 +28,16 @@ use reth_provider::BlockReaderIdExt;
 use reth_rpc_server_types::result::internal_rpc_err;
 use serde::{Deserialize, Serialize};
 use std::{collections::BTreeMap, sync::Arc, time::Duration};
-use tokio::{
-    sync::{Mutex, OnceCell},
-    time::{sleep, timeout},
-};
-use tracing::{debug, warn};
+use tokio::{sync::Mutex, time::sleep};
+use tracing::debug;
+
+mod source;
+mod state;
+
+use source::SourceLogClients;
+use state::{CachedBlock, CrossUnsafeState, SourceRef};
 
 const CROSS_UNSAFE_HEAD_POLL_INTERVAL: Duration = Duration::from_secs(5);
-/// Per-request timeout for every source-chain RPC call.
-const SOURCE_RPC_TIMEOUT: Duration = Duration::from_secs(10);
 
 /// A block head returned by `eth_crossUnsafeHead`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
@@ -470,373 +472,9 @@ const fn initiating_timestamp_in_window(
         initiating_timestamp >= executing_timestamp.saturating_sub(expiry_window)
 }
 
-#[derive(Debug)]
-struct SourceLogClients {
-    clients: Vec<SourceLogClient>,
-}
-
-impl SourceLogClients {
-    fn new(source_rpcs: impl IntoIterator<Item = impl Into<String>>) -> Result<Self, String> {
-        let mut clients: Vec<SourceLogClient> = Vec::new();
-        for source_rpc in source_rpcs {
-            let endpoint = source_rpc.into();
-            if clients.iter().any(|client| client.endpoint == endpoint) {
-                return Err(format!("duplicate source RPC URL {endpoint:?}"));
-            }
-            clients.push(SourceLogClient::new(endpoint)?);
-        }
-
-        Ok(Self { clients })
-    }
-
-    /// Finds the configured source client serving `chain_id`, resolving each client's chain ID
-    /// lazily via `eth_chainId`. A client whose RPC is not reachable yet resolves to `None` and is
-    /// simply skipped, so a peer EL that has not started cannot match (and cannot deadlock
-    /// startup); it will match on a later refresh once it is up.
-    async fn client_for(&self, chain_id: u64) -> Option<&SourceLogClient> {
-        for client in &self.clients {
-            if client.resolve_chain_id().await == Some(chain_id) {
-                return Some(client);
-            }
-        }
-        None
-    }
-
-    async fn validate_initiating_message(
-        &self,
-        chain_id: u64,
-        block_number: u64,
-        log_index: u64,
-        timestamp: u64,
-        origin: alloy_primitives::Address,
-        payload_hash: B256,
-    ) -> Option<SourceRef> {
-        let Some(client) = self.client_for(chain_id).await else {
-            warn!(
-                target: "rpc::cross_unsafe",
-                chain_id,
-                block_number,
-                log_index,
-                "no source RPC available for initiating message chain"
-            );
-            return None;
-        };
-
-        let block_hash = client
-            .validate_initiating_message(block_number, log_index, timestamp, origin, payload_hash)
-            .await?;
-        Some(SourceRef { chain_id, block_number, block_hash })
-    }
-
-    /// Returns whether the cached source block is still canonical:
-    /// - `Some(true)`  — confirmed canonical (hash matches),
-    /// - `Some(false)` — confirmed reorged (the block is reachable but its hash changed),
-    /// - `None`        — could not determine (source RPC unreachable / block not yet available), in
-    ///   which case callers must NOT treat it as a reorg.
-    async fn source_block_canonical(
-        &self,
-        chain_id: u64,
-        block_number: u64,
-        block_hash: B256,
-    ) -> Option<bool> {
-        let Some(client) = self.client_for(chain_id).await else {
-            warn!(
-                target: "rpc::cross_unsafe",
-                chain_id,
-                block_number,
-                "no source RPC available for cached source block"
-            );
-            return None;
-        };
-
-        Some(client.block_hash(block_number).await? == block_hash)
-    }
-}
-
-#[derive(Debug)]
-struct SourceLogClient {
-    endpoint: String,
-    client: RpcClient,
-    /// The source chain's ID, discovered once via `eth_chainId` and cached.
-    chain_id: OnceCell<u64>,
-}
-
-impl SourceLogClient {
-    fn new(endpoint: String) -> Result<Self, String> {
-        let url = endpoint
-            .parse()
-            .map_err(|err| format!("invalid cross unsafe head source RPC URL: {err}"))?;
-        let http = reqwest::Client::builder()
-            .timeout(SOURCE_RPC_TIMEOUT)
-            .build()
-            .map_err(|err| format!("failed to build cross unsafe head source RPC client: {err}"))?;
-        let client = ClientBuilder::default().http_with_client(http, url);
-        Ok(Self { endpoint, client, chain_id: OnceCell::new() })
-    }
-
-    /// Resolves and caches this source's chain ID via `eth_chainId`.
-    ///
-    /// Returns `None` when the RPC is not reachable yet, so a peer EL that has not started cannot
-    /// deadlock op-reth startup (no synchronous request is made when the clients are constructed).
-    /// The result is cached on first success; failures are not cached, so the next refresh retries.
-    async fn resolve_chain_id(&self) -> Option<u64> {
-        self.chain_id
-            .get_or_try_init(|| async {
-                self.fetch_chain_id().await.map(|id| id.saturating_to::<u64>()).ok_or(())
-            })
-            .await
-            .ok()
-            .copied()
-    }
-
-    /// Validates an initiating message against this source chain, returning the source block hash
-    /// on success (so the caller can record the dependency).
-    async fn validate_initiating_message(
-        &self,
-        block_number: u64,
-        log_index: u64,
-        timestamp: u64,
-        origin: alloy_primitives::Address,
-        payload_hash: B256,
-    ) -> Option<B256> {
-        let block = self.fetch_block(block_number).await?;
-        if block.header.timestamp() != timestamp {
-            debug!(
-                target: "rpc::cross_unsafe",
-                endpoint = %self.endpoint,
-                block_number,
-                expected = timestamp,
-                actual = block.header.timestamp(),
-                "source block timestamp mismatch"
-            );
-            return None;
-        }
-
-        // Filter by the claimed origin address: the message is only valid if the log at
-        // `log_index` was emitted by `origin`, so restricting to that address is a correct
-        // narrowing and keeps the response small. The block-global `log_index` is preserved by the
-        // RPC, so matching it against the filtered subset is still correct.
-        let logs = self.fetch_logs(block.header.hash, origin).await?;
-        let Some(log) = logs.into_iter().find(|log| log.log_index == Some(log_index)) else {
-            debug!(
-                target: "rpc::cross_unsafe",
-                endpoint = %self.endpoint,
-                block_number,
-                log_index,
-                "source log not found"
-            );
-            return None;
-        };
-
-        if log.address() != origin {
-            debug!(
-                target: "rpc::cross_unsafe",
-                endpoint = %self.endpoint,
-                block_number,
-                log_index,
-                expected = %origin,
-                actual = %log.address(),
-                "source log origin mismatch"
-            );
-            return None;
-        }
-
-        let remote_payload_hash = keccak256(RawMessagePayload::from(&log.inner).as_ref());
-        if remote_payload_hash != payload_hash {
-            debug!(
-                target: "rpc::cross_unsafe",
-                endpoint = %self.endpoint,
-                block_number,
-                log_index,
-                expected = %payload_hash,
-                actual = %remote_payload_hash,
-                "source log payload hash mismatch"
-            );
-            return None;
-        }
-
-        Some(block.header.hash)
-    }
-
-    /// Returns the canonical hash of `block_number` on the source chain, if available.
-    async fn block_hash(&self, block_number: u64) -> Option<B256> {
-        self.fetch_block(block_number).await.map(|block| block.header.hash)
-    }
-
-    async fn fetch_chain_id(&self) -> Option<U64> {
-        match timeout(SOURCE_RPC_TIMEOUT, self.client.request_noparams::<U64>("eth_chainId")).await
-        {
-            Ok(Ok(chain_id)) => Some(chain_id),
-            Ok(Err(err)) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    %err,
-                    endpoint = %self.endpoint,
-                    "failed to fetch source chain ID"
-                );
-                None
-            }
-            Err(_) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    endpoint = %self.endpoint,
-                    "timed out fetching source chain ID"
-                );
-                None
-            }
-        }
-    }
-
-    async fn fetch_logs(
-        &self,
-        block_hash: B256,
-        origin: alloy_primitives::Address,
-    ) -> Option<Vec<RpcLog>> {
-        let filter = Filter::new().at_block_hash(block_hash).address(origin);
-        match timeout(
-            SOURCE_RPC_TIMEOUT,
-            self.client.request::<_, Vec<RpcLog>>("eth_getLogs", (filter,)),
-        )
-        .await
-        {
-            Ok(Ok(logs)) => Some(logs),
-            Ok(Err(err)) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    %err,
-                    endpoint = %self.endpoint,
-                    %block_hash,
-                    "failed to fetch source logs"
-                );
-                None
-            }
-            Err(_) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    endpoint = %self.endpoint,
-                    %block_hash,
-                    "timed out fetching source logs"
-                );
-                None
-            }
-        }
-    }
-
-    async fn fetch_block(&self, block_number: u64) -> Option<RpcBlock> {
-        match timeout(
-            SOURCE_RPC_TIMEOUT,
-            self.client.request::<_, Option<RpcBlock>>(
-                "eth_getBlockByNumber",
-                (U64::from(block_number), false),
-            ),
-        )
-        .await
-        {
-            Ok(Ok(Some(block))) => Some(block),
-            Ok(Ok(None)) => {
-                debug!(
-                    target: "rpc::cross_unsafe",
-                    endpoint = %self.endpoint,
-                    block_number,
-                    "source block not found"
-                );
-                None
-            }
-            Ok(Err(err)) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    %err,
-                    endpoint = %self.endpoint,
-                    block_number,
-                    "failed to fetch source block"
-                );
-                None
-            }
-            Err(_) => {
-                warn!(
-                    target: "rpc::cross_unsafe",
-                    endpoint = %self.endpoint,
-                    block_number,
-                    "timed out fetching source block"
-                );
-                None
-            }
-        }
-    }
-}
-
-/// A source-chain block that a cached cross-unsafe block depends on.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct SourceRef {
-    chain_id: u64,
-    block_number: u64,
-    block_hash: B256,
-}
-
-/// In-memory cache of locally validated cross-unsafe blocks.
-///
-/// `safe` is the floor anchor; `validated` holds the contiguous chain of validated blocks strictly
-/// above it. The head is always the highest validated block, or `safe` when none are cached.
-#[derive(Debug, Default)]
-struct CrossUnsafeState {
-    safe: CachedBlock,
-    validated: BTreeMap<u64, CachedBlock>,
-}
-
-impl CrossUnsafeState {
-    fn head(&self) -> &CachedBlock {
-        self.validated.values().next_back().unwrap_or(&self.safe)
-    }
-
-    /// Re-anchors to the current safe head, dropping any cached block at or below it.
-    fn reseed_safe(&mut self, safe: CachedBlock) {
-        self.validated = self.validated.split_off(&safe.number.saturating_add(1));
-        self.safe = safe;
-    }
-
-    fn is_validated(&self, number: u64, hash: B256, parent_hash: B256) -> bool {
-        self.validated
-            .get(&number)
-            .is_some_and(|block| block.hash == hash && block.parent_hash == parent_hash)
-    }
-
-    fn insert(&mut self, block: CachedBlock) {
-        self.validated.insert(block.number, block);
-    }
-
-    /// Drops every cached block at or above `number`.
-    fn truncate_from(&mut self, number: u64) {
-        self.validated.split_off(&number);
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-struct CachedBlock {
-    number: u64,
-    hash: B256,
-    parent_hash: B256,
-    /// Distinct source blocks the executing messages in this block depend on.
-    sources: Vec<SourceRef>,
-}
-
-impl From<&CachedBlock> for CrossUnsafeHead {
-    fn from(value: &CachedBlock) -> Self {
-        Self { number: U64::from(value.number), hash: value.hash }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    fn block(number: u64, tag: u8, parent: u8) -> CachedBlock {
-        CachedBlock {
-            number,
-            hash: B256::repeat_byte(tag),
-            parent_hash: B256::repeat_byte(parent),
-            sources: Vec::new(),
-        }
-    }
 
     #[test]
     fn initiating_timestamp_window_boundaries() {
@@ -856,84 +494,5 @@ mod tests {
         // strictly-earlier initiating timestamp (down to 0) is in window.
         assert!(initiating_timestamp_in_window(0, 100, window));
         assert!(!initiating_timestamp_in_window(101, 100, window));
-    }
-
-    #[test]
-    fn builds_source_clients_from_urls() {
-        let clients = SourceLogClients::new([
-            "http://chain-a:8545".to_string(),
-            "http://chain-b:8545".to_string(),
-        ])
-        .unwrap();
-
-        assert_eq!(clients.clients.len(), 2);
-    }
-
-    #[test]
-    fn rejects_invalid_or_duplicate_source_rpc_urls() {
-        // Malformed URL.
-        assert!(SourceLogClients::new(["not a url".to_string()]).is_err());
-        // Duplicate URLs are rejected (chain IDs are not yet known at construction time).
-        assert!(
-            SourceLogClients::new([
-                "http://chain-a:8545".to_string(),
-                "http://chain-a:8545".to_string(),
-            ])
-            .is_err()
-        );
-    }
-
-    #[test]
-    fn head_defaults_to_safe_anchor() {
-        let mut state = CrossUnsafeState::default();
-        state.reseed_safe(block(5, 5, 0));
-        assert_eq!(state.head().number, 5);
-        assert_eq!(state.head().hash, B256::repeat_byte(5));
-    }
-
-    #[test]
-    fn insert_advances_head_and_truncate_rewinds_to_safe() {
-        let mut state = CrossUnsafeState::default();
-        state.reseed_safe(block(5, 5, 0));
-        state.insert(block(6, 6, 5));
-        state.insert(block(7, 7, 6));
-        assert_eq!(state.head().number, 7);
-
-        state.truncate_from(7);
-        assert_eq!(state.head().number, 6);
-
-        state.truncate_from(6);
-        assert_eq!(state.head().number, 5);
-        assert_eq!(state.head().hash, B256::repeat_byte(5));
-    }
-
-    #[test]
-    fn reseed_safe_drops_entries_at_or_below_safe() {
-        let mut state = CrossUnsafeState::default();
-        state.reseed_safe(block(5, 5, 0));
-        state.insert(block(6, 6, 5));
-        state.insert(block(7, 7, 6));
-
-        state.reseed_safe(block(6, 60, 0));
-        assert!(!state.validated.contains_key(&6));
-        assert_eq!(state.safe.number, 6);
-        assert_eq!(state.head().number, 7);
-
-        // Re-anchoring above the cached head leaves only the safe anchor.
-        state.reseed_safe(block(9, 90, 0));
-        assert!(state.validated.is_empty());
-        assert_eq!(state.head().number, 9);
-    }
-
-    #[test]
-    fn is_validated_requires_hash_and_parent_match() {
-        let mut state = CrossUnsafeState::default();
-        state.reseed_safe(block(5, 5, 0));
-        state.insert(block(6, 6, 5));
-
-        assert!(state.is_validated(6, B256::repeat_byte(6), B256::repeat_byte(5)));
-        assert!(!state.is_validated(6, B256::repeat_byte(99), B256::repeat_byte(5)));
-        assert!(!state.is_validated(6, B256::repeat_byte(6), B256::repeat_byte(99)));
-        assert!(!state.is_validated(8, B256::repeat_byte(8), B256::repeat_byte(7)));
     }
 }
